+++ src/editgrouppage.cpp
@@ -84,10 +84,10 @@ void EditGroupPage::updateDevices(std::list<cor::Light> groupDevices, std::list<
 void EditGroupPage::resize(bool resizeFullWidget) {
     QSize size = qobject_cast<QWidget*>(this->parent())->size();
     if (resizeFullWidget) {
-        this->setGeometry(size.width()  * 0.125f,
-                          size.height() * 0.125f,
-                          size.width()  * 0.75f,
-                          size.height() * 0.75f);
+        this->setGeometry(int(size.width()  * 0.125f),
+                          int(size.height() * 0.125f),
+                          int(size.width()  * 0.75f),
+                          int(size.height() * 0.75f));
     }
     mScrollAreaWidget->resize();
 }
